+++ src/global.d.ts
@@ -1,5 +1,4 @@
 declare interface Window {
   webkit: any;
-  AndroidBridge?: Record<string, undefined | ((serializedData: string) => void)>;
-  CustomEvent: typeof CustomEvent | typeof undefined;
+  AndroidBridge?: Record<string, (serializedData: string) => void>;
 }
