+++ src/promisifier.ts
@@ -1,22 +1,31 @@
-import { VKConnectSend, VKConnectSubscribeOrUnsubscribe } from './';
-import { IOMethodName, ReceiveData, ErrorData, RequestIdProp, RequestProps } from './types';
+import {
+  VKConnectSubscribeHandler,
+  RequestMethodName,
+  RequestProps,
+  RequestIdProp,
+  ReceiveData,
+  ErrorData,
+  IOMethodName
+} from './types/connect';
 
 /**
- * Creates counter interface
+ * Creates counter interface.
  */
-const createCounter = () => ({
-  current: 0,
-  next() {
-    this.current += 1;
+function createCounter() {
+  return {
+    current: 0,
+    next() {
+      this.current += 1;
 
-    return this.current;
-  }
-});
+      return this.current;
+    }
+  };
+}
 
 /**
- * Creates interface for resolving request promises by request id's (or not)
+ * Creates interface for resolving request promises by request id's (or not).
  */
-const createRequestResolver = () => {
+function createRequestResolver() {
   type PromiseController = {
     resolve: (value: any) => any;
     reject: (reason: any) => any;
@@ -27,10 +36,11 @@ const createRequestResolver = () => {
 
   return {
     /**
-     * Adds new controller with resolve/reject methods
-     * @param resolve Resolve function
-     * @param reject Reject function
-     * @returns New request id of the added controller
+     * Adds new controller with resolve/reject methods.
+     *
+     * @param resolve Resolve function.
+     * @param reject Reject function.
+     * @returns New request id of the added controller.
      */
     add: (controller: PromiseController): number => {
       const id = counter.next();
@@ -42,10 +52,11 @@ const createRequestResolver = () => {
 
     /**
      * Resolves/rejects an added promise by request id and the `isSuccess`
-     * predicate
-     * @param requestId Request ID
+     * predicate.
+     *
+     * @param requestId Request ID.
      * @param data Data to pass to the resolve- or reject-function.
-     * @param isSuccess Predicate to select the desired function
+     * @param isSuccess Predicate to select the desired function.
      */
     resolve: <T>(requestId: number | string, data: T, isSuccess: (data: T) => boolean) => {
       const requestPromise = promiseControllers[requestId];
@@ -61,14 +72,18 @@ const createRequestResolver = () => {
       }
     }
   };
-};
+}
 
 /**
- * Returns send function that returns promises
- * @param send VK Connect send method
- * @param subscribe VK Connect subscribe method
+ * Returns send function that returns promises.
+ *
+ * @param sendEvent VK Connect send event.
+ * @param subscribe VK Connect subscribe event.
  */
-export const promisifySend = (send: VKConnectSend, subscribe: VKConnectSubscribeOrUnsubscribe) => {
+export function promisifySend(
+  sendEvent: <K extends RequestMethodName>(method: K, params?: RequestProps<K> & RequestIdProp) => void,
+  subscribe: (fn: VKConnectSubscribeHandler) => void
+) {
   const requestResolver = createRequestResolver();
 
   // Subscribe to receive a data
@@ -84,13 +99,14 @@ export const promisifySend = (send: VKConnectSend, subscribe: VKConnectSubscribe
     }
   });
 
-  return <K extends IOMethodName>(method: K, props?: RequestProps<K>): Promise<ReceiveData<K>> =>
-    new Promise((resolve, reject) => {
+  return function promisifiedSend<K extends IOMethodName>(method: K, props?: RequestProps<K>): Promise<ReceiveData<K>> {
+    return new Promise((resolve, reject) => {
       const requestId = requestResolver.add({ resolve, reject });
 
-      send(method, {
+      sendEvent(method, {
         ...(props as RequestProps<K>),
         request_id: requestId
       });
     });
-};
+  };
+}
