+++ src/types/data.ts
@@ -1,9 +1,3 @@
-/** Type of global object with VK Connect methods in Android app WebView */
-export type AndroidBridge = Record<RequestMethodName, (serializedData: string) => void>;
-
-/** Type of global object with VK Connect methods in iOS app WebView */
-export type IOSBridge = Record<RequestMethodName, { postMessage?: (data: any) => void }>;
-
 /** Type of the Personal Card */
 export type PersonalCardType = 'phone' | 'email' | 'address';
 
@@ -782,106 +776,3 @@ export type ReceiveDataMap = {
   VKWebAppGyroscopeChanged: { x: string; y: string; z: string };
   VKWebAppDeviceMotionChanged: { alpha: string; beta: string; gamma: string };
 };
-
-/** Name of the method that can be sent */
-export type RequestMethodName = keyof RequestPropsMap;
-
-/** Name of the method that can be received */
-export type ReceiveMethodName = keyof ReceiveDataMap;
-
-/** Name of the method that can be only sent */
-export type RequestOnlyMethodName = Exclude<RequestMethodName, ReceiveMethodName>;
-
-/** Name of the method that can be only received */
-export type ReceiveOnlyMethodName = Exclude<ReceiveMethodName, RequestMethodName>;
-
-/** Type of any method name */
-export type MethodName = RequestMethodName | ReceiveMethodName;
-
-/** The name of the method that can be both sent and received */
-export type IOMethodName = RequestMethodName & ReceiveMethodName;
-
-/** Getter of request properties of a method */
-export type RequestProps<M extends RequestMethodName = RequestMethodName> = RequestPropsMap[M];
-
-/** Getter of response data of a method */
-export type ReceiveData<M extends ReceiveMethodName = ReceiveMethodName> = ReceiveDataMap[M];
-
-/** Property for matching sent request and received message */
-export type RequestIdProp = { request_id?: number | string };
-
-/** Client error data */
-export type ErrorDataClientError = {
-  error_code: number;
-  error_reason: string;
-  error_description?: string;
-};
-
-/** API error data */
-export type ErrorDataAPIError = {
-  error_code: number;
-  error_msg: string;
-  request_params: string[];
-};
-
-/** Auth error data */
-export type ErrorDataAuthError = {
-  error_code: number;
-  error_reason: string;
-  error_description?: string[];
-};
-
-export type ErrorData =
-  | {
-      error_type: 'client_error';
-      error_data: ErrorDataClientError;
-      request_id?: number | string;
-    }
-  | {
-      error_type: 'api_error';
-      error_data: ErrorDataAPIError;
-      request_id?: number | string;
-    }
-  | {
-      error_type: 'auth_error';
-      error_data: ErrorDataAuthError;
-      request_id?: number | string;
-    };
-
-export type VKConnectErrorEvent = {
-  detail: {
-    type: string; // TODO
-    data: ErrorData;
-  };
-};
-
-export type VKConnectSuccessEvent<T extends ReceiveMethodName> = {
-  detail: {
-    type: string; // TODO
-    data: ReceiveData<T> & RequestIdProp;
-    app_id?: string;
-    scheme?: AppearanceSchemeType;
-    appearance?: AppearanceType;
-  };
-};
-
-/** VK Connect event */
-export type VKConnectEvent<T extends ReceiveMethodName> = VKConnectErrorEvent | VKConnectSuccessEvent<T>;
-
-/** The type of function that will be subscribed to VK Connect events */
-export type VKConnectSubscribeHandler = (event: VKConnectEvent<ReceiveMethodName>) => void;
-
-/** Sending method function type */
-export type VKConnectSend = <K extends RequestMethodName>(method: K, params?: RequestProps<K> & RequestIdProp) => void;
-
-/** Subscribing/unsubscribing method function type */
-export type VKConnectSubscribeOrUnsubscribe = (fn: VKConnectSubscribeHandler) => void;
-
-/** Subscribing/unsubscribing method function type */
-export type VKConnectSupports = (method: string) => boolean;
-
-/** Type of promisified send function */
-export type VKConnectSendPromise = <K extends IOMethodName>(
-  method: K,
-  props?: RequestProps<K>
-) => Promise<ReceiveData<K>>;
